requirements.txt
Flask
python-dotenv
Flask-Login
Flask-Bcrypt
bleach